created UserProfileController file and moved get-methods to UserService
