+++ backend/src/main/java/com/lingvoFriend/backend/Services/ChatService/UserService.java
@@ -1,15 +1,16 @@
 package com.lingvoFriend.backend.Services.ChatService;
 
-import com.lingvoFriend.backend.Repositories.UserRepository;
-import com.lingvoFriend.backend.Services.AuthService.models.UserModel;
-import com.lingvoFriend.backend.Services.ChatService.models.Message;
+import java.util.List;
+import java.util.Optional;
+import java.util.StringJoiner;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.StringJoiner;
+import com.lingvoFriend.backend.Repositories.UserRepository;
+import com.lingvoFriend.backend.Services.AuthService.models.UserModel;
+import com.lingvoFriend.backend.Services.ChatService.models.Message;
 
 @Service
 public class UserService {
@@ -48,6 +49,21 @@ public class UserService {
             joiner.add(interest);
         }
         return joiner.toString();
+    }
+
+    public List<String> getGoals(String username) {
+        UserModel user = findOrThrow(username);
+        return user.getGoals(); 
+    }
+
+    public List<String> getInterests(String username) {
+        UserModel user = findOrThrow(username);
+        return user.getInterests(); 
+    }
+
+    public String getLevel(String username) {
+        UserModel user = findOrThrow(username);
+        return user.getCefrLevel(); 
     }
 
     @Autowired
